Add files via upload
I was a Data Analyst working at a Real Estate Investment Trust. The Trust would like to start investing in Residential real estate. I was tasked with determining the market price of a house given a set of features. I analyzed and predicted housing prices using attributes or features such as square footage, number of bedrooms, number of floors, and so on.